refactor: move code to fs repository

+++ internal/repository/fs_invoice.go
@@ -2,8 +2,8 @@ package repository
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
 
@@ -14,8 +14,8 @@ type FsInvoice struct {
 	basePath string
 }
 
-func NewFsInvoice(invoiceDir string) *FsInvoice {
-	basePath, err := filepath.Abs(invoiceDir)
+func NewFsInvoice(baseDir string) *FsInvoice {
+	basePath, err := filepath.Abs(baseDir)
 	if err != nil {
 		fmt.Printf("Error while getting absolute path for invoice dir. %v", err)
 		return nil
@@ -25,12 +25,6 @@ func NewFsInvoice(invoiceDir string) *FsInvoice {
 		basePath: basePath,
 	}
 }
-
-//	List() []*model.Invoice
-//	Create(invoice *model.Invoice) error
-//	Read(invoiceID string) *model.Invoice
-//	Update(invoiceID string, invoice *model.Invoice) *model.Invoice
-//	Delete(invoiceID string) error
 
 func (fi *FsInvoice) List() []*model.Invoice {
 	invoices := make([]*model.Invoice, 0)
@@ -66,12 +60,15 @@ func (fi *FsInvoice) List() []*model.Invoice {
 }
 
 func (fi *FsInvoice) Create(invoice *model.Invoice) error {
-	jsonBytes, err := json.Marshal(invoice)
+	jsonBytes, err := json.MarshalIndent(invoice, "", "  ")
 	if err != nil {
 		return err
 	}
 
 	invoicePath := filepath.Join(fi.basePath, invoice.ID+".json")
+	if checkFileExists(invoicePath) {
+		return errors.New("invoice already exists")
+	}
 
 	err = os.WriteFile(invoicePath, jsonBytes, 0o600)
 	if err != nil {
@@ -94,7 +91,7 @@ func (fi *FsInvoice) Read(invoiceID string) *model.Invoice {
 }
 
 func (fi *FsInvoice) Update(invoiceID string, invoice *model.Invoice) *model.Invoice {
-	jsonBytes, err := json.Marshal(invoice)
+	jsonBytes, err := json.MarshalIndent(invoice, "", "  ")
 	if err != nil {
 		fmt.Printf("Error while marshaling invoice %s. %v", invoiceID, err)
 		return nil
@@ -121,26 +118,4 @@ func (fi *FsInvoice) Delete(invoiceID string) error {
 	}
 
 	return nil
-}
-
-func readInvoiceFromFile(invoicePath string) (*model.Invoice, error) {
-	jsonFile, err := os.Open(invoicePath)
-	if err != nil {
-		return nil, err
-	}
-	defer jsonFile.Close()
-
-	jsonBytes, err := io.ReadAll(jsonFile)
-	if err != nil {
-		return nil, err
-	}
-
-	invoice := &model.Invoice{}
-
-	err = json.Unmarshal(jsonBytes, invoice)
-	if err != nil {
-		return nil, err
-	}
-
-	return invoice, nil
 }
